add gitlab enable to other locals language & rename GitLab folder
