+++ src/i18n/locales/de.json
@@ -114,6 +114,7 @@
     "general_template_configuration": "Allgemeine Vorlagenkonfiguration",
     "github": "GitHub",
     "github_advisories": "GitHub Advisories",
+    "gitlab": "GitLab",
     "go_modules": "Go Module",
     "group": "Gruppe",
     "hackage": "Hackage",
@@ -132,6 +133,7 @@
     "integration_defectdojo_enable": "DefectDojo-Integration aktivieren",
     "integration_defectdojo_reimport_enable": "Erneuten Import aktivieren",
     "integration_fortify_ssc_enable": "Fortify SSC-Integration aktivieren",
+    "integration_gitlab_enable": "Enable GitLab Integration",
     "integration_kenna_connector_id": "Konnektor-ID",
     "integration_kenna_enable": "Kenna Security-Integration aktivieren",
     "integrations": "Integrationen",
