+++ src/i18n/locales/en.json
@@ -109,14 +109,13 @@
     "experimental_bom_upload_v2_info": "The logic how uploaded BOMs are processed and ingested into Dependency-Track has been overhauled to be more reliable and efficient. Further, BOM processing is now an atomic operation, such that errors occurring midway do not cause a partial state to be left behind. De-duplication of components and services is more predictable, and log messages emitted during processing contain additional context, making them easier to correlate. Because the new implementation can have a big impact on how Dependency-Track behaves regarding BOM uploads, it is disabled by default for this release.",
     "experimental_info": "The following options are experimental options and should be used with precaution.",
     "fortify_ssc": "Fortify SSC",
-    "include_archived_projects": "Include archived projects",
     "gem": "Gem",
     "general": "General",
     "general_template_configuration": "General template configuration",
     "github": "GitHub",
     "github_advisories": "GitHub Advisories",
+    "gitlab": "GitLab",
     "gitlab_integration": "GitLab",
-    "integration_gitlab_enable": "Enable GitLab Integration",
     "go_modules": "Go Modules",
     "group": "Group",
     "hackage": "Hackage",
@@ -124,6 +123,7 @@
     "hours": "Hours",
     "identifier": "Identifier",
     "include_active_children": "Include active children of projects",
+    "include_archived_projects": "Include archived projects",
     "include_children": "Include children of projects",
     "index_consistency_check_cadence": "Cadence (in minutes)",
     "index_consistency_check_description": "You can enable a periodic background task that will check that all indexes exists, are not corrupted and that their difference with Dependency Track database is under a defined threshold percentage. Any check failure will trigger a rebuild of the corresponding index. A restart is required to take cadence modification into account.",
@@ -135,6 +135,7 @@
     "integration_defectdojo_enable": "Enable DefectDojo integration",
     "integration_defectdojo_reimport_enable": "Enable reimport",
     "integration_fortify_ssc_enable": "Enable Fortify SSC integration",
+    "integration_gitlab_enable": "Enable GitLab Integration",
     "integration_kenna_connector_id": "Connector ID",
     "integration_kenna_enable": "Enable Kenna Security integration",
     "integrations": "Integrations",
