+++ src/i18n/locales/es.json
@@ -114,6 +114,7 @@
     "general_template_configuration": "Configuración general de plantillas",
     "github": "GitHub",
     "github_advisories": "Avisos de GitHub",
+    "gitlab": "GitLab",
     "go_modules": "Ir a módulos",
     "group": "Grupo",
     "hackage": "Hackage",
@@ -132,6 +133,7 @@
     "integration_defectdojo_enable": "Habilitar la integración de DefectDojo",
     "integration_defectdojo_reimport_enable": "Habilitar reimportación",
     "integration_fortify_ssc_enable": "Habilite la integración de Fortify SSC",
+    "integration_gitlab_enable": "Enable GitLab Integration",
     "integration_kenna_connector_id": "ID del conector",
     "integration_kenna_enable": "Habilitar la integración de Kenna Security",
     "integrations": "Integraciones",
