+++ src/i18n/locales/fr.json
@@ -114,6 +114,7 @@
     "general_template_configuration": "Configuration générale des modèles",
     "github": "GitHub",
     "github_advisories": "GitHub Advisories",
+    "gitlab": "GitLab",
     "go_modules": "Modules Go",
     "group": "Groupe",
     "hackage": "Hackage",
@@ -132,6 +133,7 @@
     "integration_defectdojo_enable": "Activer l'intégration de DefectDojo",
     "integration_defectdojo_reimport_enable": "Activer la réimportation",
     "integration_fortify_ssc_enable": "Activer l'intégration de Fortify SSC",
+    "integration_gitlab_enable": "Enable GitLab Integration",
     "integration_kenna_connector_id": "ID du connecteur",
     "integration_kenna_enable": "Activer l'intégration de Kenna Security",
     "integrations": "Intégrations",
