+++ src/i18n/locales/hi.json
@@ -114,6 +114,7 @@
     "general_template_configuration": "सामान्य टेम्पलेट कॉन्फ़िगरेशन",
     "github": "GitHub",
     "github_advisories": "GitHub सलाह",
+    "gitlab": "GitLab",
     "go_modules": "जाओ मॉड्यूल",
     "group": "समूह",
     "hackage": "Hackage",
@@ -132,6 +133,7 @@
     "integration_defectdojo_enable": "DefectDojo एकीकरण सक्षम करें",
     "integration_defectdojo_reimport_enable": "पुनः आयात सक्षम करें",
     "integration_fortify_ssc_enable": "फोर्टिफाई एसएससी एकीकरण सक्षम करें",
+    "integration_gitlab_enable": "Enable GitLab Integration",
     "integration_kenna_connector_id": "कनेक्टर आईडी",
     "integration_kenna_enable": "केना सुरक्षा एकीकरण सक्षम करें",
     "integrations": "एकीकरण",
