+++ src/i18n/locales/it.json
@@ -114,6 +114,7 @@
     "general_template_configuration": "Configurazione generale del modello",
     "github": "GitHub",
     "github_advisories": "Avvisi su GitHub",
+    "gitlab": "GitLab",
     "go_modules": "Vai Moduli",
     "group": "Gruppo",
     "hackage": "Hackage",
@@ -132,6 +133,7 @@
     "integration_defectdojo_enable": "Abilita l'integrazione di DefectDojo",
     "integration_defectdojo_reimport_enable": "Abilita reimportazione",
     "integration_fortify_ssc_enable": "Abilita l'integrazione Fortify SSC",
+    "integration_gitlab_enable": "Enable GitLab Integration",
     "integration_kenna_connector_id": "ID del connettore",
     "integration_kenna_enable": "Abilita l'integrazione di Kenna Security",
     "integrations": "Integrazioni",
