+++ src/i18n/locales/ja.json
@@ -114,6 +114,7 @@
     "general_template_configuration": "一般的なテンプレート設定",
     "github": "GitHub",
     "github_advisories": "GitHub Advisories",
+    "gitlab": "GitLab",
     "go_modules": "Go Modules",
     "group": "グループ",
     "hackage": "Hackage",
@@ -132,6 +133,7 @@
     "integration_defectdojo_enable": "DefectDojo統合を有効にする",
     "integration_defectdojo_reimport_enable": "再インポートを有効にする",
     "integration_fortify_ssc_enable": "Fortify SSC統合を有効にする",
+    "integration_gitlab_enable": "Enable GitLab Integration",
     "integration_kenna_connector_id": "コネクタID",
     "integration_kenna_enable": "Kenna Security統合を有効にする",
     "integrations": "統合",
