+++ src/i18n/locales/pl.json
@@ -114,6 +114,7 @@
     "general_template_configuration": "Ogólna konfiguracja szablonu",
     "github": "GitHub",
     "github_advisories": "Doradztwo GitHuba",
+    "gitlab": "GitLab",
     "go_modules": "Przejdź do modułów",
     "group": "Grupa",
     "hackage": "Hackage",
@@ -132,6 +133,7 @@
     "integration_defectdojo_enable": "Włącz integrację DefectDojo",
     "integration_defectdojo_reimport_enable": "Włącz ponowny import",
     "integration_fortify_ssc_enable": "Włącz integrację Fortify SSC",
+    "integration_gitlab_enable": "Enable GitLab Integration",
     "integration_kenna_connector_id": "Identyfikator złącza",
     "integration_kenna_enable": "Włącz integrację Kenna Security",
     "integrations": "Integracje",
