+++ src/i18n/locales/pt-BR.json
@@ -114,6 +114,7 @@
     "general_template_configuration": "Configuração geral do modelo",
     "github": "GitHub",
     "github_advisories": "Avisos do GitHub",
+    "gitlab": "GitLab",
     "go_modules": "Módulos Go",
     "group": "Grupo",
     "hackage": "Hackage",
@@ -132,6 +133,7 @@
     "integration_defectdojo_enable": "Habilite a integração do DefectDojo",
     "integration_defectdojo_reimport_enable": "Ativar reimportação",
     "integration_fortify_ssc_enable": "Habilite a integração do Fortify SSC",
+    "integration_gitlab_enable": "Enable GitLab Integration",
     "integration_kenna_connector_id": "ID do conector",
     "integration_kenna_enable": "Habilite a integração do Kenna Security",
     "integrations": "Integrações",
