+++ src/i18n/locales/ru.json
@@ -114,6 +114,7 @@
     "general_template_configuration": "Общая настройка шаблонов",
     "github": "GitHub",
     "github_advisories": "Рекомендации GitHub",
+    "gitlab": "GitLab",
     "go_modules": "Go Modules",
     "group": "Группа",
     "hackage": "Hackage",
@@ -132,6 +133,7 @@
     "integration_defectdojo_enable": "Включить интеграцию с DefectDojo",
     "integration_defectdojo_reimport_enable": "Включить повторный импорт",
     "integration_fortify_ssc_enable": "Включить интеграцию с Fortify SSC",
+    "integration_gitlab_enable": "Enable GitLab Integration",
     "integration_kenna_connector_id": "Идентификатор коннектора",
     "integration_kenna_enable": "Включить интеграцию с Kenna Security",
     "integrations": "Интеграции",
