+++ src/i18n/locales/uk-UA.json
@@ -114,6 +114,7 @@
     "general_template_configuration": "Загальна конфігурація шаблону",
     "github": "GitHub",
     "github_advisories": "Повідомлення GitHub",
+    "gitlab": "GitLab",
     "go_modules": "Go Modules",
     "group": "Група",
     "hackage": "Hackage",
@@ -132,6 +133,7 @@
     "integration_defectdojo_enable": "Увімкнути інтеграцію DefectDojo",
     "integration_defectdojo_reimport_enable": "Увімкнути повторний імпорт",
     "integration_fortify_ssc_enable": "Увімкнути інтеграцію Fortify SSC",
+    "integration_gitlab_enable": "Enable GitLab Integration",
     "integration_kenna_connector_id": "Ідентифікатор конектора",
     "integration_kenna_enable": "Увімкніть інтеграцію Kenna Security",
     "integrations": "Інтеграції",
