+++ src/i18n/locales/zh.json
@@ -114,6 +114,7 @@
     "general_template_configuration": "常规模板配置",
     "github": "GitHub",
     "github_advisories": "GitHub Advisories",
+    "gitlab": "GitLab",
     "go_modules": "Go Modules",
     "group": "组",
     "hackage": "Hackage",
@@ -132,6 +133,7 @@
     "integration_defectdojo_enable": "启用 DefectDojo 集成",
     "integration_defectdojo_reimport_enable": "启用重新导入",
     "integration_fortify_ssc_enable": "启用 Fortify SSC 集成",
+    "integration_gitlab_enable": "Enable GitLab Integration",
     "integration_kenna_connector_id": "连接器 ID",
     "integration_kenna_enable": "启用 Kenna Security 集成",
     "integrations": "集成",
